Done build relationships and validate user

+++ src/main/java/vn/ITSS/teacherexchange/controller/UserController.java
@@ -2,6 +2,9 @@ package vn.ITSS.teacherexchange.controller;
 
 import java.util.List;
 
+import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,6 +15,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import vn.ITSS.teacherexchange.domain.User;
 import vn.ITSS.teacherexchange.service.UserService;
+import vn.ITSS.teacherexchange.util.annotation.ApiMessage;
+import vn.ITSS.teacherexchange.util.error.IdInvalidException;
 
 @RestController
 public class UserController {
@@ -22,17 +27,27 @@ public class UserController {
     }
 
     @PostMapping("/user/create")
-    public User createNewUser(@RequestBody User postManUser) {
-
-        User ericUser = this.userService.handleCreateUser(postManUser);
-
-        return ericUser;
+    @ApiMessage("Create a new user")
+    public ResponseEntity<User> createNewUser(@Valid @RequestBody User postManUser) throws IdInvalidException {
+        boolean isEmailExist = this.userService.isEmailExist(postManUser.getEmail());
+        if (isEmailExist) {
+            throw new IdInvalidException(
+                    "Email " + postManUser.getEmail() + " đã tồn tại, vui lòng sử dụng email khác.");
+        }
+        User khoaUser = this.userService.handleCreateUser(postManUser);
+        return ResponseEntity.status(HttpStatus.CREATED).body(khoaUser);
     }
 
     @DeleteMapping("/user/delete/{id}")
-    public String deleteUser(@PathVariable("id") long id) {
+    @ApiMessage("Delete a user")
+    public ResponseEntity<Void> deleteUser(@PathVariable("id") long id) throws IdInvalidException {
+        User currentUser = this.userService.fetchUserById(id);
+        if (currentUser == null) {
+            throw new IdInvalidException("User với id = " + id + " không tồn tại");
+        }
+
         this.userService.handleDeleteUser(id);
-        return "ericUser";
+        return ResponseEntity.ok(null);
     }
 
     // fetch user by id
@@ -48,9 +63,13 @@ public class UserController {
     }
 
     @PutMapping("/user/update")
-    public User updateUser(@RequestBody User user) {
-        User ericUser = this.userService.handleUpdateUser(user);
-        return ericUser;
+    @ApiMessage("Update a user")
+    public ResponseEntity<User> updateUser(@RequestBody User user) throws IdInvalidException{
+        User khoaUser = this.userService.handleUpdateUser(user);
+        if (khoaUser == null) {
+            throw new IdInvalidException("User với id = " + user.getId() + " không tồn tại");
+        }
+        return ResponseEntity.ok(khoaUser);
     }
 
 }
